📦 feat(refactor): reorganize schemas and service implementations
